+++ src/FakeId.php
@@ -8,14 +8,14 @@ namespace byrokrat\id;
  * Fake personal identity numbers
  *
  * Fake ids replace serial number post delimiter with xxxx. If sex should be
- * encoded xx1x or xx2x can be used.
+ * encoded xxFx, xxMx or xxOx can be used, denoting Female, Male or Other.
  */
 class FakeId extends PersonalId
 {
     /**
      * Regular expression describing id structure
      */
-    const PATTERN = '/^((?:\d\d)?)(\d{6})([-+]?)(xx[12x])(x)$/i';
+    const PATTERN = '/^((?:\d\d)?)(\d{6})([-+]?)(xx[0-9xfmo])(x)$/i';
 
     public function __construct(string $number)
     {
@@ -27,7 +27,13 @@ class FakeId extends PersonalId
 
     public function getSex(): string
     {
-        return is_numeric($this->getSerialPostDelimiter()[2]) ? parent::getSex() : Sexes::SEX_UNDEFINED;
+        foreach ([Sexes::SEX_FEMALE, Sexes::SEX_MALE, Sexes::SEX_OTHER, Sexes::SEX_UNDEFINED] as $sexIdentifier) {
+            if (strcasecmp($sexIdentifier, $this->getSerialPostDelimiter()[2]) === 0) {
+                return $sexIdentifier;
+            }
+        }
+
+        return parent::getSex();
     }
 
     public function getBirthCounty(): string