+++ tests/FakeIdTest.php
@@ -104,25 +104,60 @@ class FakeIdTest extends TestCase
         );
     }
 
-    public function testGetSex()
+    public function sexProvider(): array
+    {
+        return [
+            ['770314-xxFx', Sexes::SEX_FEMALE],
+            ['770314-xxfx', Sexes::SEX_FEMALE],
+            ['770314-xx2x', Sexes::SEX_FEMALE],
+            ['770314-xx4x', Sexes::SEX_FEMALE],
+            ['770314-xxMx', Sexes::SEX_MALE],
+            ['770314-xxmx', Sexes::SEX_MALE],
+            ['770314-xx1x', Sexes::SEX_MALE],
+            ['770314-xx3x', Sexes::SEX_MALE],
+            ['770314-xxOx', Sexes::SEX_OTHER],
+            ['770314-xxox', Sexes::SEX_OTHER],
+            ['770314-xxXx', Sexes::SEX_UNDEFINED],
+            ['770314-xxxx', Sexes::SEX_UNDEFINED],
+        ];
+    }
+
+    /**
+     * @dataProvider sexProvider
+     */
+    public function testGetSex(string $raw, string $expextedSex): void
+    {
+        $this->assertEquals(
+            $expextedSex,
+            (new FakeId($raw))->getSex()
+        );
+    }
+
+    public function testSexInspectorMethods()
     {
         $fakeId = new FakeId('820323-xx1x');
-        $this->assertEquals(Sexes::SEX_MALE, $fakeId->getSex());
         $this->assertTrue($fakeId->isMale());
         $this->assertFalse($fakeId->isFemale());
+        $this->assertFalse($fakeId->isSexOther());
         $this->assertFalse($fakeId->isSexUndefined());
 
         $fakeId = new FakeId('770314-xx2x');
-        $this->assertEquals(Sexes::SEX_FEMALE, $fakeId->getSex());
         $this->assertFalse($fakeId->isMale());
         $this->assertTrue($fakeId->isFemale());
+        $this->assertFalse($fakeId->isSexOther());
         $this->assertFalse($fakeId->isSexUndefined());
 
         $fakeId = new FakeId('770314-xxxx');
-        $this->assertEquals(Sexes::SEX_UNDEFINED, $fakeId->getSex());
         $this->assertFalse($fakeId->isMale());
         $this->assertFalse($fakeId->isFemale());
+        $this->assertFalse($fakeId->isSexOther());
         $this->assertTrue($fakeId->isSexUndefined());
+
+        $fakeId = new FakeId('770314-xxOx');
+        $this->assertFalse($fakeId->isMale());
+        $this->assertFalse($fakeId->isFemale());
+        $this->assertTrue($fakeId->isSexOther());
+        $this->assertFalse($fakeId->isSexUndefined());
     }
 
     public function testGetBirthCounty()